feat(backend cloud): fix dbpass and conn_str remove azure log handler
