+++ src/backend/sfm/config.py
@@ -8,37 +8,6 @@ import psycopg2
 load_dotenv()
 
 
-def generate_db_string(ENV: str, DBHOST: str, DBNAME: str, DBUSER: str, DBPASS: str):
-    """Take in env variables and generate correct db string."""
-
-    if ENV == "test":
-        return "sqlite://"  # in-memory database for unit tests
-
-    if ENV == "local":
-        return "sqlite:///./issues.db"  # local sqlite for local development
-
-    if ENV == "development" or "production":
-        # need all four parameters available
-        if "unset" in [DBNAME, DBPASS]:
-            raise ValueError(
-                "Missing database parameter in the environment.  Please specify DBHOST, DBNAME, DBUSER, and DBPASS"
-            )
-
-        # driver = "{ODBC Driver 17 for SQL Server}"
-        # conn = f"""dbname='{DBNAME}' user='sfadmin@psql-sfm' host='psql-sfm.postgres.database.usgovcloudapi.net'
-        # password='{DBPASS}' port='5432' sslmode='true'"""
-        conn = "host={0} user={1} dbname={2} password={3} sslmode={4}".format(
-            DBHOST, DBUSER, DBNAME, DBPASS, "require"
-        )
-        # conn = f"""Driver={driver};Server=tcp:{DBHOST},1433;Database={DBNAME};
-        # Uid={DBUSER};Pwd={DBPASS};Encrypt=yes;TrustServerCertificate=no;Connection Timeout=30;"""
-        # params = urllib.parse.quote_plus(conn)
-        conn_str = psycopg2.connect(conn)
-
-        # return conn
-        return conn_str
-
-
 class Settings(BaseSettings):
     """
     App settings class.  These values should be exposed as environment variables.  If running locally, this can be a `.env` file (see included `.env.example` file).  If running in another enviornment you should use a modern approach to manage secrets.  Many of these values are secrets and should not be exposed in the source code or to a user of the application.  Many settings and decisions are based on the `ENV` variable's value.
@@ -72,7 +41,9 @@ class Settings(BaseSettings):
     DBUSER: str = os.environ.get("DBUSER") or "unset"
     DBPASS: str = os.environ.get("DBPASS") or "unset"
     AZURE_LOGGING_CONN_STR: str = os.environ.get("AZURE_LOGGING_CONN_STR") or "unset"
-    CONN_STR: str = generate_db_string(ENV, DBHOST, DBNAME, DBUSER, DBPASS)
+
+
+#     CONN_STR: str = generate_db_string(ENV, DBHOST, DBNAME, DBUSER, DBPASS)
 
 
 @lru_cache()
