+++ src/backend/sfm/routes/metrics/routes.py
@@ -23,9 +23,9 @@ logging.basicConfig(
 )
 
 logger = logging.getLogger(__name__)
-logger.addHandler(
-    AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
-)
+# logger.addHandler(
+#     AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
+# )
 
 router = APIRouter()
