+++ src/backend/sfm/routes/projects/crud.py
@@ -20,9 +20,9 @@ logging.basicConfig(
 
 logger = logging.getLogger(__name__)
 
-logger.addHandler(
-    AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
-)
+# logger.addHandler(
+#     AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
+# )
 
 
 def get_all(db: Session, skip: int = None, limit: int = None):
