+++ src/backend/sfm/routes/work_items/crud.py
@@ -17,9 +17,9 @@ logging.basicConfig(
 
 logger = logging.getLogger(__name__)
 
-logger.addHandler(
-    AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
-)
+# logger.addHandler(
+#     AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
+# )
 
 
 def get_all(
