+++ src/backend/sfm/routes/work_items/routes.py
@@ -20,9 +20,9 @@ logger = logging.getLogger(__name__)
 
 router = APIRouter()
 
-logger.addHandler(
-    AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
-)
+# logger.addHandler(
+#     AzureLogHandler(connection_string=app_settings.AZURE_LOGGING_CONN_STR)
+# )
 
 
 @router.get("/", response_model=List[WorkItemRead])
